# docs/openapi-mock.yaml
openapi: 3.1.0
info:
  title: 차곡차곡 API v0.2 (Mock 명세)
  version: 0.2.0-mock
  description: |
    빌라 골목 주차 공유 앱 **차곡차곡**의 API 명세(v0.2)입니다. 프런트가 서버 없이 화면을 만들 수 있도록 **예시 값은 모두 가짜(Mock) 데이터**로 채웠습니다.

    ### 기준 문서
    - **Figma "월계디버깅 / 와이어프레임" 페이지가 기준**입니다 (문구·선택지·화면 흐름). Manyfast 와이어프레임 23개 화면은 같은 화면의 이전본이고, 요구사항(PRD)은 Manyfast 를 참고합니다.
    - Notion "차곡차곡 API 명세 v0.2" 는 **임시 API 명세**입니다 (엔드포인트·필드·에러 코드 초안). Figma 와 다르면 Figma 를 따릅니다. 화면 번호 #1~#23 은 Notion 의 "화면 → API 매핑" 번호와 같습니다.
    - **DB 에 저장할 곳이 없는 기능은 API 로 만들지 않습니다** (알림 설정, Web Push 구독, 로그아웃, 공유 요청 취소, 미확인 차량 메모, 긴급 이동 요청 등).

    - 현재 DB 스키마 (`app/models/`). 아래 네 가지는 Notion v0.2 대신 **DB 를 따릅니다**.

    ### Notion v0.2 와 다르게 정한 것 (DB 기준)
    | 항목 | Notion v0.2 | 이 명세 (DB 기준) |
    |---|---|---|
    | 결제 | 원 | **토큰**. `residents.token_balance` 로 보유, 수락 시 요청자 → 공유를 연 관리인에게 이동 (`token_transfers`) |
    | 차고지(Garage) | 관리인이 등록하는 공유 상품 (칸 묶음 + 시간 + 요금) | **빌라의 주차장**. 빌라 하나에 차고지 하나라서 API 의 `garage_id` 는 빌라 id 와 같음. 시간·토큰 가격은 칸마다 따로 두는 **공유 조건** (`share_offers`) 에 있음 |
    | 계층 | Building > Zone > Lane > Slot (`depth`) | **골목(Alley) > 빌라(Building) = 차고지 > 주차 구역(Zone) > 칸(Slot)**. 주차 구역은 필로티 안쪽·건물 앞·골목 같은 묶음 (DB 테이블 이름은 `garages`). 막힘은 `front_slot_id` 로 계산 |
    | 사용자–빌라 | 여러 빌라 소속 가능 (`BuildingMember`) | **한 사용자는 한 빌라에만 소속** (`residents.building_id`). 이미 소속된 사용자가 합류하면 409 |

...

    ### 공통 규칙
    | 항목 | 규칙 |
    |---|---|
    | 인증 | `Authorization: Bearer <access_token>` (`🔓` 표시 엔드포인트는 불필요) |
...
    | 금액 | 정수(**토큰**). 원 결제는 없음 |
...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /buildings/{building_id}/vehicle-reports:
    post:
      tags: [vehicle-reports]
      summary: 👤 미등록 차량 사진 제보
      description: |
        앱에 없는 차를 사진으로 제보합니다. **한 트랜잭션**으로 아래가 함께 일어납니다.
        - 그 칸에 미확인 차량(`occupant_type=UNKNOWN`) 배치 → 배치도·현황·막힘 판정에 바로 반영
        - 빌라의 모든 관리인(제보자 본인 제외)에게 `VEHICLE_REPORT` 알림 (`link: { screen: VEHICLE_REPORT, id: 제보 id }`)
        - 제보자에게 **500 토큰** 시스템 지급
        사진: JPEG·PNG, 10MB 이하. 서버가 EXIF(위치정보 등)를 지우고 긴 변 1600px JPEG 로 다시 저장합니다.
        검증 순서: 403 `NOT_BUILDING_MEMBER` → 400 `INVALID_INPUT`(`detail.field` = `photo` | `plate`) → 409 `REPORT_LIMIT_EXCEEDED`(1인 하루(KST) 3회) →
        404(이 빌라에 없는 칸) → 409 `SLOT_UNAVAILABLE`(비활성 칸) · `SLOT_OCCUPIED`(차 있음) · `PLATE_EXISTS`(앱에 등록된 차 → 이동 요청 안내) · `VEHICLE_ALREADY_PARKED`(같은 번호판 미확인 차량이 다른 칸에 있음)
      operationId: createVehicleReport
      x-priority: 향후
      parameters:
        - { $ref: "#/components/parameters/BuildingId" }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo, plate, slot_id]
              properties:
                photo: { type: string, format: binary, description: "차량 사진 (JPEG·PNG, 10MB 이하)" }
                plate: { type: string, description: "차량 번호. 공백 유무 모두 허용", example: 45다 6789 }
                slot_id: { type: integer, example: 1006 }
      responses:
        "201":
          description: 제보됨
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VehicleReportCreated" }
              example: { report_id: 12, vehicle_id: 31, parking_id: 559, slot_label: P6, occupant_type: UNKNOWN, reward_tokens: 500, token_balance: 500500 }
        "400": { $ref: "#/components/responses/InvalidInput" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotBuildingMember" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: 칸·번호판·하루 한도
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                limit: { summary: 하루 한도 초과, value: { error: { code: REPORT_LIMIT_EXCEEDED, message: 오늘은 더 제보할 수 없어요., detail: { limit: 3 } } } }
                plateExists: { summary: 앱에 등록된 차, value: { error: { code: PLATE_EXISTS, message: 이미 등록된 차량 번호입니다., detail: null } } }
  /vehicle-reports/{report_id}:
    get:
      tags: [vehicle-reports]
      summary: 👤 제보 상세 (제보자 본인·그 빌라 관리인)
      description: "알림(`VEHICLE_REPORT`) 탭 → 이 화면. 볼 수 없는 사람에게는 404 입니다."
      operationId: getVehicleReport
      x-priority: 향후
      parameters:
        - { $ref: "#/components/parameters/ReportId" }
      responses:
        "200":
          description: 제보 상세
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VehicleReportDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /vehicle-reports/{report_id}/photo:
    get:
      tags: [vehicle-reports]
      summary: 👤 제보 사진 (제보자 본인·그 빌라 관리인)
      description: "번호판이 찍힌 개인정보라 공개 URL 이 없습니다. Bearer 인증으로 받아 Blob 으로 그립니다. `Cache-Control: private`"
      operationId: getVehicleReportPhoto
      x-priority: 향후
      parameters:
        - { $ref: "#/components/parameters/ReportId" }
      responses:
        "200":
          description: 사진 (JPEG)
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/buildings/{building_id}/vehicle-reports:
    get:
      tags: [admin, vehicle-reports]
      summary: 🛠 제보 목록 (최신순)
      operationId: listVehicleReports
      x-priority: 향후
      parameters:
        - { $ref: "#/components/parameters/BuildingId" }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
      responses:
        "200":
          description: 제보 목록
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VehicleReportPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/NotBuildingAdmin" }

  # ───────────────────────── 알림 ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BuildingId: { name: building_id, in: path, required: true, schema: { type: integer }, example: 3 }
...
    ReportId: { name: report_id, in: path, required: true, schema: { type: integer }, example: 12 }
...
    Cursor: { name: cursor, in: query, required: false, description: 이전 응답의 next_cursor, schema: { type: string } }
    Limit: { name: limit, in: query, required: false, schema: { type: integer, default: 20, minimum: 1, maximum: 50 } }
  responses:
    InvalidInput:
      description: 검증 실패
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: INVALID_INPUT, message: 입력값이 올바르지 않습니다., detail: { field: agree_terms, reason: 이용약관에 동의해야 합니다. } } }
    Unauthorized:
      description: 토큰 없음·만료
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: UNAUTHORIZED, message: 로그인이 필요합니다., detail: null } }
    NotBuildingMember:
      description: 해당 빌라 입주민이 아님
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: NOT_BUILDING_MEMBER, message: 이 빌라의 입주민만 이용할 수 있습니다., detail: null } }
    NotBuildingAdmin:
      description: 해당 빌라 관리인이 아님
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: NOT_BUILDING_ADMIN, message: 이 빌라의 관리인만 이용할 수 있습니다., detail: null } }
    NotFound:
      description: 대상을 찾을 수 없음
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: NOT_FOUND, message: 대상을 찾을 수 없습니다., detail: null } }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message: { type: string }
            detail: { type: [object, "null"], additionalProperties: true }
    ErrorCode:
      type: string
      description: |
        400 `INVALID_INPUT` · 401 `UNAUTHORIZED` `INVALID_CREDENTIALS` · 403 `NOT_BUILDING_MEMBER` `NOT_BUILDING_ADMIN` ·
        404 `NOT_FOUND` `INVALID_INVITE_CODE` · 409 `CONFLICT`(범용, 현재 상태와 충돌) `EMAIL_EXISTS` `SLOT_OCCUPIED` `SLOT_UNAVAILABLE` `VEHICLE_ALREADY_PARKED` `GARAGE_TIME_CONFLICT`.
        **제안(v0.2 미정의)**: `PLATE_EXISTS` `ALREADY_DECIDED` `MOVE_REQUEST_ALREADY_PENDING` `ALREADY_IN_BUILDING` `INSUFFICIENT_TOKENS` `REPORT_LIMIT_EXCEEDED`(미등록 차량 제보 하루 한도, #52) (모두 409)
      enum:
        - INVALID_INPUT
        - UNAUTHORIZED
        - INVALID_CREDENTIALS
        - NOT_BUILDING_MEMBER
        - NOT_BUILDING_ADMIN
        - NOT_FOUND
        - INVALID_INVITE_CODE
        - CONFLICT
        - EMAIL_EXISTS
        - SLOT_OCCUPIED
        - SLOT_UNAVAILABLE
        - VEHICLE_ALREADY_PARKED
        - GARAGE_TIME_CONFLICT
        - PLATE_EXISTS
        - ALREADY_DECIDED
        - MOVE_REQUEST_ALREADY_PENDING
        - ALREADY_IN_BUILDING
        - INSUFFICIENT_TOKENS
        - REPORT_LIMIT_EXCEEDED
...
    OccupantType:
      type: string
      enum: [RESIDENT, EXTERNAL, UNKNOWN]
      description: 입주민 차량 / 외부·공유 이용자 / 미확인
...
    VehicleReportCreated:
      type: object
      properties:
        report_id: { type: integer }
        vehicle_id: { type: integer }
        parking_id: { type: integer }
        slot_label: { type: string }
        occupant_type: { $ref: "#/components/schemas/OccupantType" }
        reward_tokens: { type: integer, description: 지급한 보상 토큰 (500) }
        token_balance: { type: integer, description: 보상 지급 후 내 잔액 }
    VehicleReportDetail:
      type: object
      properties:
        id: { type: integer }
        building_id: { type: integer }
        plate: { type: string, example: 45다 6789 }
        slot_label: { type: string }
        photo_url: { type: string, description: "API 기준 경로 (/vehicle-reports/{id}/photo). Bearer 인증 필요", example: /vehicle-reports/12/photo }
        reporter:
          type: object
          properties:
            id: { type: integer }
            name: { type: string, description: 이름. 없으면 닉네임 }
        status: { type: string, enum: [SUBMITTED, DISMISSED], description: "DISMISSED = 관리인이 허위 제보로 기각 (기각 API 는 향후)" }
        created_at: { type: string, format: date-time }
    VehicleReportPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/VehicleReportDetail" }
        next_cursor: { type: [string, "null"] }
...